pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lensprint"
version = "0.1.0"
description = "Transform photos with blurred borders and EXIF overlays."
readme = "README.md"
authors = [
    { name = "Yuan Chen", email = "yuan.chen@offerplus.io" }
]
license = { file = "LICENSE" }
requires-python = ">=3.7"
dependencies = [
    "pillow",
    "opencv-python",
    "piexif"
]
keywords = ["photo", "exif", "border", "blur", "photography"]

[project.urls]
Homepage = "https://github.com/chenyuan99/lensprint"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.0.0"
]

[tool.ruff]
line-length = 88
target-version = "py37"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
